requirements.txt

attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0

certifi==2021.5.30
django-filter==2.4.0
cffi==1.14.6
channels==3.0.4
channels-redis==3.3.0
charset-normalizer==2.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.6
django-nested-admin==3.3.3
django-redis==5.0.0
djoser==2.0.1
djangochannelsrestframework==0.3.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
drf-friendly-errors==0.14
hiredis==2.0.0
hyperlink==21.0.0
idna==3.2
incremental==21.3.0
kombu==5.1.0
mixins==0.1.4
msgpack==1.0.2
prompt-toolkit==3.0.20
mysqlclient==2.0.3
djangorestframework-datatables==0.6.0
future==0.18.2
timezones==2.0.9

